chore: 快照 Nuxt AiGate，准备 All-in Go 重建
Preserve the full TypeScript/Nuxt platform state (gateway, org/quota, MCP, KB, agents, skills) as the legacy baseline prior to greenfield Go rewrite.

+++ server/api/aigate/billing/[id].get.ts
@@ -37,7 +37,7 @@ export default defineEventHandler(async event => {
         model: apiLog.model,
         requests: sql<number>`count(*)::int`,
         tokens: sql<number>`coalesce(sum(${apiLog.totalTokens}), 0)::int`,
-        cost: sql<number>`coalesce(sum(${apiLog.cost}), 0)::int`,
+        cost: sql<number>`coalesce(sum(${apiLog.cost}), 0)`,
       })
       .from(apiLog)
       .where(and(...logConditions))
@@ -49,7 +49,7 @@ export default defineEventHandler(async event => {
         date: sql<string>`to_char(${apiLog.createdAt}, 'YYYY-MM-DD')`,
         requests: sql<number>`count(*)::int`,
         tokens: sql<number>`coalesce(sum(${apiLog.totalTokens}), 0)::int`,
-        cost: sql<number>`coalesce(sum(${apiLog.cost}), 0)::int`,
+        cost: sql<number>`coalesce(sum(${apiLog.cost}), 0)`,
       })
       .from(apiLog)
       .where(and(...logConditions))
